GEMINI_FINETUNE/requirements.txt
# Core dependencies
pandas
numpy
scikit-learn
openpyxl

# LLM & API
google-generativeai
vertexai
python-dotenv
requests

# Document parsing
python-docx
pdfplumber

# Utils
mimetypes2